chore: moved request-password to auth routes
